rename _helpers, remove print statements
